data_pipelines/scripts/transcription_requirements.txt
# Core dependencies for the customer journey analysis pipeline
pandas==2.2.2
numpy==1.26.4
openai==1.30.1
mistralai==0.4.0
python-dotenv==1.0.1
torchaudio
transformers
scipy
pyannote.audio
scikit-learn
google-generativeai
python-dotenv
mistralai>=0.4.0
openai
tiktoken
tqdm